Turns out the tests aren't passing.
Need to sort the dump method in the model, There's an example of what it
should look like in the test package. It's just sorting out how it
prints. That would help if anyone can sort it out. The game is also
grinding to a halt. It might be the randomNumberGenerator.